Fix map not showing any content if the route is invalid
In any situation where we couldn't fetch directions for a route, the map
would just not show up. This patch fixes that by providing an initial
center and zoom to the map, which it will use until the
DirectionsRenderer attaches.

We still need to actually report to the user what went wrong with their
route (too complex, too many waypoints, unresolvable, etc.).